+++ src/components/Tooltip/__stories__/TooltipContent.jsx
@@ -2,7 +2,7 @@ import React from "react";
 import "./TooltipContent.scss";
 
 const TooltipContent = () => {
-  return <div>Tooltip Content</div>;
+  return <div>Text</div>;
 };
 
 export default TooltipContent;